requirements.txt
streamlit
pandas
pytrends
openai
python-dotenv
